Updated the Vector class to fix a bug with the operator=, continued work on the renderer. Now there are buffers and whilst they don't do anything they will soon enough. Comments explain the whole situation so don't goob it up future me.

+++ BbCore/src/structures/Vector2/Vector2f.h
@@ -10,7 +10,7 @@ namespace Bb {
 		Vector2f(const Vector2f& other);
 		Vector2f(Vector2f&& other) = delete;
 
-		float x, y;
+		unsigned int x, y;
 
 		// ASSIGNMENT
